feat: grid item page
